requirements.txt
google-adk
python-dotenv
google-generativeai
ShopifyAPI